Update functions has been implemented.
