+++ src/components/LayoutTools/index.tsx
@@ -1,0 +1,1 @@
+export { ManualLayoutPanel } from './LayoutToolsPanel'
