--- src/components/ManualLayout/index.tsx
@@ -1,1 +1,0 @@
-export { ManualLayoutPanel } from './ManualLayoutPanel'
